phase2/backend/requirements.txt
# FastAPI and ASGI Server
fastapi>=0.110.0
uvicorn>=0.27.0

# Database and ORM
sqlmodel>=0.0.14
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.0

# Authentication and JWT
PyJWT>=2.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Type hints
typing-extensions>=4.9.0

# Password hashing (bcrypt must be <5.0.0 for passlib compatibility)
passlib>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Development tools (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

